feat: implement static alarms
BREAKING CHANGE: This removes the `anomaly_detection_band_threshold`
variable.

This commit removes anomaly detection alarms and
implements static alarming thresholds on load balancer 5xx
responses. This is due to the way AWS handles the 5xx count metric.
It is impossible to build a valid model off this metric when 5xx
responses are uncommon, because this metric is a counter. When the
counter is not incremented, it is not represented as a "zero" but
as the lack of data, and thus the anomaly detection modelling
refuses to build a model.